js/src/xpconnect/tests/mochitest/test_bug564330.html


<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=564330
-->
<head>
  <title>Test for Bug 560351</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=564330">Mozilla Bug 564330</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 564330 **/
var pass = false;
try {
    document.createElement("span").setUserData("x", <y/>, null);
} catch (e) {
    pass = true;
}

</script>
<script>

ok(pass, "pass was set to true");

</script>
</pre>
</body>
</html>
